requirements.txt
certifi==2018.8.24
numpy==1.15.1
pgzero==1.2
psutil==5.4.7
pygame==1.9.4
Pyganim==0.9.2
pyserial==3.4
wincertstore==0.2
